util: split types into multiple files
