+++ packages/xchain-util/src/types/asset.ts
@@ -1,0 +1,7 @@
+import { Chain } from '../chain'
+
+export type Asset = {
+  chain: Chain
+  symbol: string
+  ticker: string
+}
